requirements/tests.txt
Flask-Babel
Babel
Flask-Login
Flask-Mailman
Flask-Mongoengine
Flask-Principal
peewee
Flask-SQLAlchemy
argon2_cffi
authlib
bcrypt
bleach
check-manifest
coverage
cryptography
python-dateutil
mongoengine
mongomock
msgcheck
passlib
pony
phonenumberslite
pydocstyle
pytest-cache
pytest-cov
pytest
qrcode
sqlalchemy
sqlalchemy-utils
webauthn>=1.6.0;python_version>='3.8'
werkzeug
zxcvbn
